refactor: Customer Repository 를 port/adaptor 구조로 분리
