+++ admin/src/main/java/kernel360/ckt/admin/infra/jpa/CustomerJpaRepository.java
@@ -1,9 +1,30 @@
 package kernel360.ckt.admin.infra.jpa;
 
 import kernel360.ckt.core.domain.entity.CustomerEntity;
-import kernel360.ckt.admin.infra.CustomerRepository;
+import kernel360.ckt.core.domain.enums.CustomerStatus;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
-public interface CustomerJpaRepository extends JpaRepository<CustomerEntity, Long>, CustomerRepository {
+import java.util.Optional;
 
+public interface CustomerJpaRepository extends JpaRepository<CustomerEntity, Long> {
+    @Query("""
+        SELECT c FROM CustomerEntity c
+        WHERE (:status IS NULL OR c.status = :status)
+          AND (
+            :keyword IS NULL
+            OR c.customerName LIKE %:keyword%
+            OR c.phoneNumber LIKE %:keyword%
+          )
+    """)
+    Page<CustomerEntity> findAll(
+        @Param("status") CustomerStatus status,
+        @Param("keyword") String keyword,
+        Pageable pageable
+    );
+
+    Optional<CustomerEntity> findByLicenseNumber(String licenseNumber);
 }
